Initial commit of libical and libecal-1.2 bindings
libical is simply not bound in the Vala world, and so these bindings
(from the Maya project) are a step toward access.  They appear to have
been generated by vapigen, which is sketchy because libical isn't a
GObject library.

libecal-1.2 is bound because the default bindings distributed by
Vala use void pointers when referring to libical structures (which are
unavailable because libical isn't bound; see above).  With the
libical bindings, a better but-not-perfect binding is possible.  The
.metadata file for libecal-1.2 was also borrowed from the Maya
project.