Display recurring events: Closes bgo#725788
Recurring events are now handled properly internally, although
creating/updating/removing will have to wait.  Thus, an instance
of a recurring event can only be examined now, cannot be updated.

This also fixes a bug where the month view only displayed events
within the month but not on days outside of the month that were
in view.