Make main window resizable under Unity: Closes bgo#726705
Unity under Trusty 14.04 simply does not allow the application window
to be resized if a GtkHeaderBar is set as its titlebar.  No
combination of related sizing properties for GtkWindow seems to
change this.

So, for Unity, the --enable-unity configure switch must be used.
This means the main window's decorations will remain.