Hide/show calendars: Closes bgo#725780
This introduces the Calendar Manager, which will be expanded later
to allow the user to add, remove, and configure their calendars.
Right now the only active control in it is to set the visibility
of the calendars.  It's also tied to the EDS setting, meaning
changing visibility in another EDS client (i.e. Evolution) will
be reflected in California.