Update current date when date changes: Closes bgo#726139
Detecting if the current date has changed is done through two
mechanisms: (a) if the timezone changes and that change results in
a date change, and (b) polling the system clock occassionally for
a date change.  The second feels wasteful but there's no system
event I can locate that will inform an app when the date has changed
and don't want to poll too infrequently in case the user/NTP updates
the clock manually.