Create "Toolkit" namespace for widgetry
This also abolishes the old clunky Host.Interaction and makes all
UI panes Toolkit.Cards hosted in Toolkit.Decks.  This is a more
flexible system and allows for these cards to be strung together
in single-window dialogs as well as GtkPopovers.

Finally, the old color chooser popup was removed in favor of using
GtkColorButton in the Calendar Manager.