Week View: Closes bug #725767
Introduces Week View, which required refactoring the host container
(main window) and refactoring a number of month view widgets to share
code between the two views.  Although a large patch, this sets up
California to more easily add other views and in general operate in a
consistent way in all of them.

A number of simplifications were also introduced in the toolkit/ unit,
which can be used in future code and potentially backported to
existing code.