Fix Quick Add bug parsing "12:30pm"
Internally time is represented by a 24-hour clock and the parser
was adding 12 to 12pm to convert.  Fixed and tests added.

Added conversion of "2400" to midnight.

In quick parser, if no date is specified but midnight is, it
assumes the event is midnight of the next day.  Might generalize this
logic to always use next day if no day specified and time is in the
current day's past.