Improve Show Event dialog
Show Event used to be a scrolling text label with all the details
inserted as markup.  It wasn't very well organized and didn't look
very pleasing.

This new design isn't a radical improvement but it does make the
pertinent details (where, when) more obvious and easier to locate.

This patch also enables selection throughout the dialog, closing
bug #730394.