Don't pass non-null Value messages holding null
It's a pain for the Toolkit.Cards to have to check for non-null
Value messages holding null references, so Toolkit.Deck will "strip"
those messages out and simply pass null to the Card.

This occurred because apparently either GSignal or some Vala-generated
code was turning a null Value into a non-null Value holding null and
the code on the other end of the signal wasn't expecting that.