Remove recurring events: Bug #725787
User can now remove an instance, all instances, or an instance and
all its future instances of a recurring event.

This also includes some fixes to the EDS CalendarSourceSubscription
class.  Investigating this ticket showed some problems with how we
were dealing with signals from the EDS CalClientView that needed to
be corrected in order to properly deal with particular instances of
a recurring event going away.