Improved time/date widgets in Create/Update event dialog: Bug #725783
This introduces new time/date widgets when creating or editing an
event.  Rather than popping up over the existing window, a new card
in the deck is shown (much like when editing recurring events)
allowing the user to visually adjust date and time with mouse and/or
the keyboard.

This also removes the sustained duration auto-adjustment (bug #732031)