Don't adjust event duration in date/time editor while typing
Last commit had one serious shortcoming: if the user was typing a
multidigit hour, the maintained duration could swing wildly.  (For
example, typing a "1" then a "2" for 12 could make a 1-hour duration
become a 12-hour duration.)  This corrects that by only adjusting the
duration if the focus has left the GtkEntry.