Allow for non-date-specified Quick Add: Bug #735939
Unity/Ambience theming bug led us not to show any GtkPopovers that
are attached to the GtkHeaderBar, instead using the current day for
Ctrl+N (or the Quick Add button).  That regressed functionality by
not allowing for users to enter a Quick Add event devoid of a
starting date.  This restores that functionality, using a DeckWindow
instead of a DeckPopover on Unity/Ambiance.