Don't set end date for single-date Quick Add: Bug #735940
If the user Quick Adds a recurring event by clicking on a date,
prior code would use that date as the start and end date of the event,
which screwed up the RRULE.  Now a single day is only treated as
the starting date of the recurring rule.

Additionally, if a date is selected and a weekly recurring event is
Quick Added that starts on a different day of the week, the original
start date is dropped (for consistency).  In other words, clicking on
a Wednesday and entering "Meeting every Tuesday" will generate an
RRULE that starts on the upcoming Tuesday.