docs: Updates for the different ways you can open the quick-add
Updated docs to include the different ways you can open the
quick-add feature. If you open the quick-add feature with the plus
or ctrl-n, you also need to include the date/time in your details.
If you double-click the desired date/time, you don't need to
include the date/time when typing the event details.