Stop main window from "stuttering" when maximized: Bug #729029
The GtkStack transition for the month and week views were causing the
main window to shift strangely on the screen and then return to
position.  Toolkit.StackModel was using the timer to "guess" when the
GtkStack transition was completed to balance its cache.  With GTK+
3.12, a GtkStack property notification does this better, and moving
the balance call to there seems to have solved the issue.