Properly move event between calendars: Bug #739164
Component.Instance allowed for callers to directly change the
Backing.CalendarSource of an instance.  This caused comparison and
hash problems when the event was to be removed from the view
collections.  Now that Instance clones are used in editing, it's not
necessary to allow changing the calendar source. Instead, the calendar
source is simply used to create/update/delete the clone as appropriate
and the subscription manager will report changes to the view.